Split diff viewer row rendering
